- Overhauled profile/entry version handling. Until now, the program would refuse to load profiles/entries of older versions at all.
Now, profiles/entries are loaded according to their actual version specified in the loaded file. Then, the profile/entry is compared to a 'best' version format, specified in the program.
If the best format contains any values which are not present in the loaded (older) format, they are filled with defaults, an information pops up and a backup of the profile is created.
This way, profile updating should now occur automatically. It has not been tested yet, though.

If actual version and best version are the same, there should be no difference to the old behavior.

- Fixed a small bug with total playtime display, where the '<' sign needed a HTML-style escape as '&lt;' in QLabel.